+++ History.md
@@ -9,6 +9,7 @@
 - It can be set up so that external links and PDF-links will be opened in new, blank browser tabs/windows.
 - Security: Consultation admins could grant privileges to other consultations within the same site.
 - The backlinkt on the "my account" page leads to the consultation where the user was coming from.
+- Admins can now download Excel/XLSX-exports of amendments and the comments to motions.
 - Bugfix: If a motion section was removed from a motion type that already had motions, then amendments to this motion could not be merged individually anymore.
 - Bugfix: If a voting block was deleted with motions/amendments assigned to it, then these motions/amendments could not be assigned to new voting blocks anymore.
 - Bugfix: If a motion had an empty optional section, amendments adding text to that section were breaking the motion view.